+++ tests/unit/test_unit.lua
@@ -13,6 +13,7 @@ local modules = {
   "activity_ui_spec",
   "query_result_renderer_spec",
   "result_filetype_spec",
+  "result_column_navigation_spec",
   "result_session_spec",
   "result_sticky_header_spec",
   "query_selection_spec",